Remove references to the neutron:server:plugin parameter
This parameter is present in service- and system-level metadata
even though it became oblosete somewhere around adoption of Kilo
release of OpenStack when it was replased by the
neutron:server:backend:engine param. Its presense currently only
clutters metadata and the formula and confuses users.

This change removes references to this param from the formula and
fixes the 2 remaining cases of its usage:
  - In Juno neutron-server conf replaced it with the `backend.engine`
  - In map.jinja removed the long contrail-related if-block. All
    metadata that is used for dict population there is provided in
    the system level classes.

Change-Id: Id8b2123d2d3deb89caa8fa12bc6608d032c00494