Refactor mechanism driver metadata
Using a quite heavy nested dict structure for defining a list of
mechanism driver to load is an overkill. This change simplifies
the structure by turning it into a list.
The mechanism driver meta for gateway and compute nodes is now moved
to the system level (like it is already done for control nodes).
The l2population mechanism driver is still enabled by default,
however its enablement is no longer unconditional.

Change-Id: I4ae9c9c0562ef2271bf627d53d62aca60c101dab