Fixing inaccurate physnet mapping iterations
In order to handle situations, when external vlan ranges are not
applicable, we should iterate properly to generate config file.
Removing physnet1 hardcode in such case.

Pillar:

  neutron:
    server:
      external_access

is not set by default anywhere, so it would be True and this
patch does not break backward compatibility for physnet1 hardcode.

Adding ironic physnet handling

Customer-Found
Change-Id: Iead3b5e2ecfa072801b6e0766751e371e36e895f